Make corners local to calibration canvas
