+++ app/_lib/key-code.ts
@@ -1,4 +1,5 @@
 export enum KeyCode {
  KeyM = 'KeyM',
  Escape = 'Escape',
+ Tab = 'Tab',
 }
